media: mc: entity: Rework media_pipeline_start() to cache pads
The media_pipeline_start() and media_pipeline_stop() functions use the
media graph walk API to traverse the graph and validate the pipeline.
While the pipeline can't change between the start and stop calls, the
graph is walked again from scratch at stop time.

This wastes CPU cycles, which can be avoided by caching pads in the
media_pipeline structure at start time, and reusing them at stop time.

Additionally, caching pads in the pipeline can serve in the future as
the fundation to provide a better API than the media graph walk to
drivers to iterate over pads and entities in the pipeline.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>