[HACK] arm64: dts: imx8mp: Add BSP-specific clocks and resets for ISI
The media-blk-ctrl handles resets and clocks differently between the
mainline and the BSP implementations of the corresponding drivers. In
the mainline kernel the clocks and resets are handled internally by the
media-blk-ctrl driver in power domains, while in the BSP driver they are
exposed to the consumers and have to be handled manually. Add the
BSP-specific clocks and resets to the ISI DT node.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>