arm64: dts: freescale: imx8mm-phyboard: Add overlay for VM-016 camera module
Add an overlay for the PHYTEC PHYCAM VM-016 camera module connected to
the i.MX8MM Polis development board. The camera is based on an ONSemi
AR0144 sensor, with an on-board EEPROM and clock oscillator. Sensor
power supplies are generated on-board from an always-on single supply
provided on the camera module connector.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>