dt-bindings: media: Add Vision Components MIPI camera module controller
The MIPI camera module controller is an FPGA-based, I2C-accessible
system controller found on MIPI camera modules from Vision Components.
It abstracts the camera sensor behind a unified register interface, and
controls and sequences the on-board power supplies and clocks.

The camera sensor abstraction is optional, the controller allows all
sensor registers to be directly written from the host. This binding
models the clock and power supply functions of the controller.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>