media: imx-mipi-csis: Add VC-related register fields
The CMN_CTRL and ISPCFG registers have fiels related to virtual channel
handling. In CMN_CTRL, the INTERLEAVE_MODE field is 2 bits wide, and
controls filtering by DT, VC or both. The VC number is then set in the
ISPCFG.VIRTUAL_CHANNEL field.

Expand the definition of the register macros to support those features,
and set the VC ID to 0 explicitly instead of relying on the default
register value. This prepares for VC filtering but does not modify the
driver's behaviour.

While at it, use GENMASK in the last few register mask macros that don't
use it.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>